Test filesystem api
* Invariant tests for the File interface
* Invariant tests for reading / writing files
** Due to missing features currently ignoring CryptoFileSystem

+++ main/filesystem-api/src/main/java/org/cryptomator/filesystem/File.java
@@ -77,7 +77,16 @@ public interface File extends Node, Comparable<File> {
 		Mover.move(this, destination);
 	}
 
+	/**
+	 * <p>
+	 * Deletes the file if it exists.
+	 * <p>
+	 * Does nothign if the file does not exist.
+	 */
 	default void delete() {
+		if (!exists()) {
+			return;
+		}
 		try (WritableFile writableFile = openWritable()) {
 			writableFile.delete();
 		}
